+++ scripts/timers.js
@@ -2,7 +2,7 @@
 
 class Timers {
 	constructor() {
-		['cursor', 'jsearch1', 'jsearch2', 'tt'].forEach(v => this[v] = {
+		['cursor', 'jsearch1', 'jsearch2', 'jsearch3', 'tt'].forEach(v => this[v] = { // Regorxxx <- Fixed quick-search on same letter
 			id: null
 		});
 	}